+++ app/Wellbeing/Assets.xcassets/BAC Icon.imageset/Contents.json
@@ -4,6 +4,14 @@
       "idiom" : "universal",
       "filename" : "BAC Icon.png",
       "scale" : "1x"
+    },
+    {
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "idiom" : "universal",
+      "scale" : "3x"
     }
   ],
   "info" : {
