+++ app/Wellbeing/EmergencyViewController.swift
@@ -19,9 +19,11 @@ class EmergencyViewController: UIViewController {
     let RICE_X = 29.719565
     let RICE_Y = -95.402233
     let RICE_RADIUS = 1000
+    let PATH = "http://0.0.0.0:5000/api/blue_button_location"
     var allowActions = false
     
     let locationService = LocationService.sharedInstance
+    let backendHandler = BackendHandler.sharedInstance
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -48,7 +50,8 @@ class EmergencyViewController: UIViewController {
     */
     @IBAction func activateEmergency(sender: AnyObject) {
         if allowActions {
-            updateLocationToServer(locationService.currentLocation!)
+            backendHandler.postLocationToServer(locationService.currentLocation!, path: PATH)
+//            updateLocationToServer(locationService.currentLocation!)
         } else {
             print("wrong place bud")
         }
@@ -68,60 +71,60 @@ class EmergencyViewController: UIViewController {
     /*
         Updates location to server
     */
-    func updateLocationToServer(location: CLLocation) {
-        
-        print("serving data")
-        
-        let caseID = NSUUID().UUIDString
-        let deviceID = UIDevice.currentDevice().identifierForVendor!.UUIDString
-        let latitude = "\(location.coordinate.latitude)"
-        let longitude = "\(location.coordinate.longitude)"
-        
-        let dateFormatter = NSDateFormatter()
-        dateFormatter.dateFormat = "MM/dd/yyyy - hh:mm:ss"
-        
-        let timestamp = dateFormatter.stringFromDate(location.timestamp)
-        
-        let postString = "caseID=" + validateURLString(caseID) + "&deviceID=" + validateURLString(deviceID) + "&longitude=" + validateURLString(longitude) + "&latitude=" + validateURLString(latitude) + "&date=" + validateURLString(timestamp) + "&resolved=false"
-        print(postString)
-        let path: String = "http://0.0.0.0:5000/api/blue_button_location"
-        let url: NSURL = NSURL(string: path)!
-        let cachePolicy = NSURLRequestCachePolicy.ReloadIgnoringLocalCacheData
-        let request = NSMutableURLRequest(URL: url, cachePolicy: cachePolicy, timeoutInterval: 2.0)
-        request.HTTPMethod = "POST"
-        request.HTTPBody = postString.dataUsingEncoding(NSUTF8StringEncoding);
-        
-        let task = NSURLSession.sharedSession().dataTaskWithRequest(request) {
-            
-            data, response, error in
-            
-            if error != nil {
-                print("error =\(error)")
-                return
-            }
-            
-            let responseString = NSString(data: data!, encoding:NSUTF8StringEncoding)
-            print("response =\(responseString!)")
-            
-            do {
-                let json = try NSJSONSerialization.JSONObjectWithData(data!, options: [.MutableContainers, .AllowFragments]) as? NSDictionary
-                if let parseJSON = json {
-                    let result = parseJSON["status"] as? String
-                    print("status =\(result)")
-                }
-                
-            } catch {
-                print("json error: \(error)")
-            }
-        }
-        
-        task.resume()
-        
-    }
-    
-    func validateURLString(string: String) -> String {
-        return string.stringByAddingPercentEncodingWithAllowedCharacters(.URLHostAllowedCharacterSet())!
-    }
+//    func updateLocationToServer(location: CLLocation) {
+//        
+//        print("serving data")
+//        
+//        let caseID = NSUUID().UUIDString
+//        let deviceID = UIDevice.currentDevice().identifierForVendor!.UUIDString
+//        let latitude = "\(location.coordinate.latitude)"
+//        let longitude = "\(location.coordinate.longitude)"
+//        
+//        let dateFormatter = NSDateFormatter()
+//        dateFormatter.dateFormat = "MM/dd/yyyy - hh:mm:ss"
+//        
+//        let timestamp = dateFormatter.stringFromDate(location.timestamp)
+//        
+//        let postString = "caseID=" + validateURLString(caseID) + "&deviceID=" + validateURLString(deviceID) + "&longitude=" + validateURLString(longitude) + "&latitude=" + validateURLString(latitude) + "&date=" + validateURLString(timestamp) + "&resolved=false"
+//        print(postString)
+//        let path: String = "http://0.0.0.0:5000/api/blue_button_location"
+//        let url: NSURL = NSURL(string: path)!
+//        let cachePolicy = NSURLRequestCachePolicy.ReloadIgnoringLocalCacheData
+//        let request = NSMutableURLRequest(URL: url, cachePolicy: cachePolicy, timeoutInterval: 2.0)
+//        request.HTTPMethod = "POST"
+//        request.HTTPBody = postString.dataUsingEncoding(NSUTF8StringEncoding);
+//        
+//        let task = NSURLSession.sharedSession().dataTaskWithRequest(request) {
+//            
+//            data, response, error in
+//            
+//            if error != nil {
+//                print("error =\(error)")
+//                return
+//            }
+//            
+//            let responseString = NSString(data: data!, encoding:NSUTF8StringEncoding)
+//            print("response =\(responseString!)")
+//            
+//            do {
+//                let json = try NSJSONSerialization.JSONObjectWithData(data!, options: [.MutableContainers, .AllowFragments]) as? NSDictionary
+//                if let parseJSON = json {
+//                    let result = parseJSON["status"] as? String
+//                    print("status =\(result)")
+//                }
+//                
+//            } catch {
+//                print("json error: \(error)")
+//            }
+//        }
+//        
+//        task.resume()
+//        
+//    }
+//    
+//    func validateURLString(string: String) -> String {
+//        return string.stringByAddingPercentEncodingWithAllowedCharacters(.URLHostAllowedCharacterSet())!
+//    }
 
     
     /*
